requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.10.4
pydantic-settings==2.7.0
python-dotenv==1.0.1
supabase==2.10.0
groq==0.13.1
httpx==0.27.2
aiohttp==3.11.11
twilio==9.4.0
apscheduler==3.10.4
pydub==0.25.1
pandas==2.2.3
numpy==1.26.4
streamlit==1.41.1
plotly==5.24.1
loguru==0.7.3